+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), ~~and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)~~.
 
+## [Unreleased]
+### Fixed
+ - Fix the link to organisation budgets from the summary screen ([#90](https://github.com/pwyf/iati-decipher/issues/90))
+
 ## [1.2.0] - 2018-11-12
 ### Changed
  - Unload all datasets, instead of animating between them